pyproject.toml
[project]
name = "guidebook-prj-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bigquery>=0.0.30",
    "dotenv>=0.9.9",
    "google-api-python-client>=1.7.11",
    "google-cloud>=0.34.0",
    "gspread>=6.2.1",
    "numpy>=2.4.3",
    "oauth2client>=4.1.3",
    "oauthlib>=3.3.1",
    "pandas>=3.0.1",
    "tqdm>=4.67.3",
]
